refactor(cli): move the daemon autostart retry seam out of client.ts
client.ts sat at the 700-line CLI source budget; withAutostart and its
stale-daemon wait are one self-contained seam (connection failure or
build drift → autostart → retry), so they get their own module rather
than the wrapper being squeezed to fit.
